# OpenAPI/API_documentation.yaml
openapi: 3.0.0
info:
  title: "DiscoverTrento API"
  description: "documentazione API per la webapp DiscoverTrento"
  version: "1.0"

#IMPORTANTE! DOBBIAMO SOSTITUIRE I SERVER CON I NOSTRI
servers:
  - url: http://localhost:3000/api/v1
    description: Localhost
components:
  securitySchemes:
    TokenAutenticazione:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
      description: "Token fornito all'utente all'atto del login necessario per accedere alle risorse riservate"
paths:
...
  /eventi:
    get:
      summary: "Lista degli eventi"
      description: "Restituisce la lista degli eventi, filtrabili per tipologia o data"
      tags: 
        - "evento"
      parameters:
        - name: "tipologia"
          in: "query"
          schema:
            enum: ["sportivo", "culturale", "volontariato"]
          description: "Tipologia dell'evento"
          example: "sportivo"
        - name: "data"
          in: "query"
          schema:
            type: "string"
            format: "date"
          description: "Data dell'evento (formato: YYYY-MM-DD)"
        - name: "luogo"
          in: "query"
          schema:
            type: "string"
          description: "Luogo dell'evento"
          example: "Piazza Duomo"
      responses:
        "200":
          description: "Lista degli eventi"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  type: "object"
                  properties:
                    id_evento:
                      type: "string"
                    nome_evento:
                      type: "string"
                    data:
                      type: "string"
                      format: "date"
                    luogo:
                      type: "string"
                    tipologia:
                      enum: ["sportivo", "culturale", "volontariato"]
                    descrizione:
                      type: "string"
                    punti:
                      type: "integer"
                    creatore:
                      type: "string"
                      description: "id del cittadino o dell'operatore comunale che ha creato l'evento"
        "500":
          description: "Errore del server"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  error:
                    type: "string"
                    example: "Errore del server, riprova più tardi"
    post:
      summary: "Proposta evento"
      description: "Consente a cittadino e operatore comunale di proporre un evento, specificandone i dettagli, luogo e data"
      tags: 
        - "evento"
      security:
        - TokenAutenticazione: []       
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                nome_evento:
                  type: "string"
                data:
                  type: "string"
                  format: "date"
                luogo:
                  type: "string"
                tipologia:
                  enum: ["sportivo", "culturale", "volontariato"]
                descrizione:
                  type: "string"
                punti:
                  type: "integer"
                  description: "Considerato solo nel caso in cui sia l'utente operatore_comunale a creare l'evento"
              required: ["nome_evento", "data", "luogo", "tipologia"]
      responses:
        "201":
          description: "La creazione dell'evento è avvenuta con successo"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  id_evento:
                    type: "string"
                    description: "ID dell'evento creato"
                    example: "12345"
                  messaggio:
                    type: "string"
                    description: "Messaggio di conferma"
                    example: "L'evento è stato creato con successo."
        "400":
          description: "Richiesta non valida: dati mancanti o non validi"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  error:
                    type: "string"
                    example: "Richiesta non valida"
        "403":   # se l'utente azienda prova a proporre un evento
          description: "Forbidden: la tipologia di utente non permette la proposta di eventi"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  error:
                    type: "string"
                    example: "Azione non permessa"
        "500":
          description: "Errore del server"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  error:
                    type: "string"
                    example: "Errore del server, riprova più tardi"
  /eventi/:id_evento:
    put: 
      summary: "Assegnazione punti ad un evento"
      description: "Permette all'utente operatore_comunale di assegnare un punteggio ad un evento"
      tags:
        - "evento"
      security:
        - TokenAutenticazione: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                punti:
                  type: "integer"
                  description: "Punteggio assegnato al coupon"
              required: [ "punti"]
      responses:
        "200":
          description: "Il punteggio è stato assegnato all'evento dal comune con successo"
        "400":
          description: "Richiesta non valida: dati mancanti o non validi"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  error:
                    type: "string"
                    example: "Richiesta non valida"
        "403":  
          description: "Forbidden: la tipologia di utente non permette la proposta di coupon"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  error:
                    type: "string"
                    example: "Azione non permessa"
        "404":
          description: "Evento non trovato"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  error:
                    type: "string"
                    example: "Evento non trovato"
        "500":
          description: "Errore del server"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  error:
                    type: "string"
                    example: "Errore del server, riprova più tardi"
...